Revert "For class variables, lookup type in base classes (#1338, #2022, #2211)"
